Move task spawn to factory
